backend/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-core-v3==2.13.33
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==37.0.4
Deprecated==1.2.13
Fiona==1.8.21
geopandas==0.11.1
idna==3.3
jmespath==0.10.0
munch==2.5.0
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pycodestyle==2.9.1
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-futures==1.0.0
Shapely==1.8.4
six==1.16.0
toml==0.10.2
urllib3==1.26.11
wrapt==1.14.1
